+++ app/src/main/res/layout/activity_home_fragment.xml
@@ -6,16 +6,9 @@
     android:layout_height="match_parent"
     tools:context=".HomeFragment">
 
-    <!-- Expandable List View -->
-    <ExpandableListView
+    <include
         android:id="@+id/expandableListView"
-        android:layout_width="0dp"
-        android:layout_height="580dp"
-        style="@style/ListStyle"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"/>
+        layout="@layout/fragment_custom_expandable_list" />
 
     <!-- Admin Button (Top Right Corner) -->
     <Button
